+++ src/PSRule.Rules.Azure/Data/ProviderDataExtensions.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Rules.Azure.Data
+{
+    /// <summary>
+    /// Defines extension methods for Azure resource provider data.
+    /// </summary>
+    internal static class ProviderDataExtensions
+    {
+        public static bool TryResourceType(this ProviderData data, string resourceType, out ResourceProviderType type)
+        {
+            type = null;
+            return ResourceHelper.TryResourceProviderFromType(resourceType, out var provider, out var typeName) &&
+                data.TryResourceType(provider, typeName, out type);
+        }
+    }
+}
