+++ src/PSRule.Rules.Azure/Data/Template/DeploymentSymbolKind.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Rules.Azure.Data.Template
+{
+    internal enum DeploymentSymbolKind
+    {
+        None = 0,
+        Array = 1,
+        Object = 2
+    }
+}
