+++ src/PSRule.Rules.Azure/Data/Template/TypePrimitive.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace PSRule.Rules.Azure.Data.Template
+{
+    [JsonConverter(typeof(StringEnumConverter))]
+    internal enum TypePrimitive
+    {
+        None,
+
+        Array,
+
+        Bool,
+
+        Int,
+
+        Object,
+
+        String,
+
+        SecureString,
+
+        SecureObject
+    }
+}
