+++ src/PSRule.Rules.Azure/Data/Template/ValidationKind.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Rules.Azure.Data.Template
+{
+    internal enum ValidationKind
+    {
+        None = 0,
+
+        Parameter = 1
+    }
+}
