+++ src/PSRule.Rules.Azure/Pipeline/ExportDataPipelineBuilder.cs
@@ -9,17 +9,6 @@ namespace PSRule.Rules.Azure.Pipeline
     public delegate AccessToken GetAccessTokenFn(string tenantId);
 
     /// <summary>
-    /// A builder to configure the pipeline to export data.
-    /// </summary>
-    public interface IExportDataPipelineBuilder : IPipelineBuilder
-    {
-        /// <summary>
-        /// Configures a method to request an Access Token.
-        /// </summary>
-        void AccessToken(GetAccessTokenFn fn);
-    }
-
-    /// <summary>
     /// A helper to build a pipeline to export data from Azure.
     /// </summary>
     internal abstract class ExportDataPipelineBuilder : PipelineBuilderBase, IExportDataPipelineBuilder
