+++ src/PSRule.Rules.Azure/Pipeline/PipelineWriter.cs
@@ -2,26 +2,12 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
 using System.Management.Automation;
 using System.Threading;
 using PSRule.Rules.Azure.Configuration;
 
 namespace PSRule.Rules.Azure.Pipeline
 {
-    internal interface ILogger
-    {
-        void WriteVerbose(string message);
-
-        void WriteVerbose(string format, params object[] args);
-
-        void WriteWarning(string message);
-
-        void WriteWarning(string format, params object[] args);
-
-        void WriteError(Exception exception, string errorId, ErrorCategory errorCategory, object targetObject);
-    }
-
     internal abstract class PipelineWriter : ILogger
     {
         protected const string ErrorPreference = "ErrorActionPreference";
@@ -175,37 +161,5 @@ namespace PSRule.Rules.Azure.Pipeline
         {
             return (ActionPreference)sessionState.PSVariable.GetValue(variableName);
         }
-    }
-
-    internal abstract class SerializationOutputWriter<T> : PipelineWriter
-    {
-        private readonly List<T> _Result;
-
-        protected SerializationOutputWriter(PipelineWriter inner, PSRuleOption option)
-            : base(inner, option)
-        {
-            _Result = new List<T>();
-        }
-
-        public override void WriteObject(object sendToPipeline, bool enumerateCollection)
-        {
-            Add(sendToPipeline);
-        }
-
-        protected void Add(object o)
-        {
-            if (o is T[] collection)
-                _Result.AddRange(collection);
-            else if (o is T item)
-                _Result.Add(item);
-        }
-
-        public sealed override void End()
-        {
-            var results = _Result.ToArray();
-            base.WriteObject(Serialize(results), false);
-        }
-
-        protected abstract string Serialize(T[] o);
     }
 }
