+++ src/PSRule.Rules.Azure/Pipeline/PolicyAssignmentPipelineBuilder.cs
@@ -7,24 +7,6 @@ using PSRule.Rules.Azure.Pipeline.Output;
 
 namespace PSRule.Rules.Azure.Pipeline
 {
-    /// <summary>
-    /// A builder for a policy assignment export pipeline.
-    /// </summary>
-    public interface IPolicyAssignmentPipelineBuilder : IPipelineBuilder
-    {
-        /// <summary>
-        /// Enable pass-through of output from the pipeline.
-        /// </summary>
-        /// <param name="passThru">Enable pass-through.</param>
-        void PassThru(bool passThru);
-
-        /// <summary>
-        /// Determines if policy definitions that duplicate existing built-in rules are exported.
-        /// </summary>
-        /// <param name="keepDuplicates">Enable exporting duplicates.</param>
-        void KeepDuplicates(bool keepDuplicates);
-    }
-
     internal sealed class PolicyAssignmentPipelineBuilder : PipelineBuilderBase, IPolicyAssignmentPipelineBuilder
     {
         private bool _PassThru;
