+++ src/PSRule.Rules.Azure/Pipeline/SerializationOutputWriter.cs
@@ -1,0 +1,40 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Collections.Generic;
+using PSRule.Rules.Azure.Configuration;
+
+namespace PSRule.Rules.Azure.Pipeline
+{
+    internal abstract class SerializationOutputWriter<T> : PipelineWriter
+    {
+        private readonly List<T> _Result;
+
+        protected SerializationOutputWriter(PipelineWriter inner, PSRuleOption option)
+            : base(inner, option)
+        {
+            _Result = new List<T>();
+        }
+
+        public override void WriteObject(object sendToPipeline, bool enumerateCollection)
+        {
+            Add(sendToPipeline);
+        }
+
+        protected void Add(object o)
+        {
+            if (o is T[] collection)
+                _Result.AddRange(collection);
+            else if (o is T item)
+                _Result.Add(item);
+        }
+
+        public sealed override void End()
+        {
+            var results = _Result.ToArray();
+            base.WriteObject(Serialize(results), false);
+        }
+
+        protected abstract string Serialize(T[] o);
+    }
+}
